requirements.txt
Brotli==1.0.7
click==7.1.2
dash==1.16.1
dash-bootstrap-components==0.10.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
DateTime==4.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==0.25.0
#pandas==1.0.3
patsy==0.5.1
plotly==4.7.1
plotly-express==0.4.1
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
scipy==1.4.1
six==1.14.0
statsmodels==0.11.1
Werkzeug==1.0.1
zope.interface==5.1.0
sqlalchemy==1.3.9
mysql-connector-python==8.0.20
psutil==5.6.3
mysqlclient==2.0.1
xlrd==1.1.0